Move unit tests into `tests/unit/`
