+++ tests/unit/tts/test_tts.py
@@ -3,7 +3,7 @@ from unittest.mock import Mock, call
 
 from parameterized import parameterized
 
-from tests.utils import assert_called_with_exactly, build_audio
+from unit.utils import assert_called_with_exactly, build_audio
 from voicebox.tts import TTS, FallbackTTS, RetryTTS
 
 log = Mock()
